+++ FintrakBanking.Repositories/WorkFlow/Workflow.cs
@@ -55,7 +55,7 @@ namespace FintrakBanking.Repositories.WorkFlow
         private bool keepPending = false;
         private bool politicallyExposed = false;
         private bool deferredExecution = false;
-        private int? vote = null;
+        private short? vote = null;
         private int? toStaffId = null;
 
         public int StaffId { set { staffId = value; } }
@@ -68,7 +68,7 @@ namespace FintrakBanking.Repositories.WorkFlow
         public int Tenor { set { tenor = value; } }
         public bool InvestmentGrade { set { investmentGrade = value; } }
         public bool PoliticallyExposed { set { politicallyExposed = value; } }
-        public int? Vote { set { vote = value; } }
+        public short? Vote { set { vote = value; } }
         public int StatusId { get { return statusId; } set { statusId = value; } }
         public int? NextLevelId { get { return nextLevelId; } set { nextLevelId = value; } }
         public int? ProductId { set { productId = value; } }
@@ -157,7 +157,7 @@ namespace FintrakBanking.Repositories.WorkFlow
                 APPROVALSTATUSID = (short)this.statusId,
                 SYSTEMARRIVALDATETIME = this.systemDate,
                 SYSTEMRESPONSEDATETIME = this.systemDate,
-                VOTE = (short)this.vote,
+                VOTE = this.vote,
                 TOSTAFFID = this.toStaffId,
             };
