+++ FintrakBanking.ViewModels/Credit/AppraisalMemorandumViewModel.cs
@@ -34,7 +34,7 @@ namespace FintrakBanking.ViewModels.Credit
         public int? receiverStaffId { get; set; }
         public decimal amount { get; set; }
         public bool politicallyExposed { get; set; }
-        public int? vote { get; set; }
+        public short? vote { get; set; }
         public string comment { get; set; }
         public decimal principal { get; set; }
         public double rate { get; set; }
